add ata derivation to solana package
