+++ packages/solana/package.json
@@ -48,6 +48,7 @@
     "typecheck": "tsc -p tsconfig.typecheck.json"
   },
   "dependencies": {
+    "@solana/spl-token": "^0.4.8",
     "@solana/web3.js": "^1.88.1",
     "@turnkey/http": "workspace:*",
     "@turnkey/sdk-browser": "workspace:*",
